// src/dto/slider.dto.ts
/**
 * Created by anjum on 30/10/16.
 */
export class SliderData {
    active: string;
    addressId: number;
    addressDescription:string;
    description: string;
    distance: number;
    distanceDenom: string;
    earnBurn: string;
    img: string;
    isCheckedIn: boolean;
    isFavourite: boolean;
    isSlidActive: boolean;
    isTarget: boolean;
    lastCheckinTime: string;
    latitude: number;
    location: string;
    locationId: number;
    locationCount: number;
    logo: string;
    longitude: number;
    myTarget: number;
    name: string;
    offerId: number;
    offerPoints: number;
    offers: string;
    offersCount: number;
    organizationDescription: string;
    pageTitle: string;
    partnerCode: string;
    points: number;
    showHide: boolean;
    special: string;
    storeName: string;
    targetEndDate: string;
    targetEndDateString: string;
    targetId: number;
    targetStatus: string;
    targetStatusDescription: string;
    terms: string;
    termsType: string;
    title: string;
}